refactor : divide like apis #48
